Factor out %host-start.r, use PARSE for command-line arg processing
This deletes the C-based command-line argument processing, in favor
of using PARSE in Rebol code.  It also makes that code exclusive to
the host linked against the Rebol interpreter...meaning that the core
doesn't do any command-line processing at all.

In order to accomplish this, the basic initialization in Init_Core()
was simplified significantly.  A pre-existing script for generating
the embedded host code (%make-host-init.r) was resurrected into
working condition, and RL_Start() removed from the API...instead
delegating it to the main() function that initialized Rebol to run
its own embedded code for startup.

Among other changes, this eliminates the "system/options/flags",
instead relying on specific system/options/xxx to tell the condition
of a setting.