Use qsort_r argument to remove global state from sort
Presumably sorting for blocks used to use the version of qsort() with
no tunneled-through parameter, and was later changed to use qsort_r,
because there was a note about not using global state.

This does that change, as well as taking advantage of the built-in
type-checking in variadic apply for the custom sorting.  (In R3-Alpha
there was no such checking, so the sort command did it itself.)