Eliminate most unused var/param warnings, action refinement errors
Though Ren-C initially started out with the goal of having the warnings
turned all the way up, that goal had fallen some by the wayside.  This
addresses some of the problems in high warning levels, along with
making progress on having no unused variable warnings.

Because actions are starting to include the argument macros, this
revealed several cases where parameters to actions were being dropped
on the floor.  As that code evolves into a new design to replace
actions, this makes some progress on giving errors when a refinement
is not going to be heeded.

(Note: This may be too conservative, if there are refinements that
are "ok to ignore" for some types... e.g. /ONLY for strings in append)