Replace EXTERNAL series with HANDLE + LENGTH
When the FFI was added by Atronix, it was able to "image" a structure
schema on top of memory that was not owned by Rebol, and then poke and
peek values out of it.  This was used for reading data created by
a non-Rebol DLL without having to copy its memory into a Rebol
series first.

However, there was also a desire to have Rebol provide the memory
backing the structures.  So a special kind of series was added which
could have its memory not controlled by Rebol, or an "external series".

While it may be tempting to consider this as a general feature of the
system, the notion that all code has to be bulletproofed generically
against the idea of not owning a series pointer is daunting.  That
adds a lot of edge cases for which the system is not prepared--not to
mention a consistent processing "tax" on knowing whether series data
is external or not.

There's now another approach to the same problem--namely that the
data which backs a struct could be *either* a BINARY! (owned by Rebol)
or a managed HANDLE! (arbitrary memory).  This pushes the
responsibility for checking if data is external over to the STRUCT!
implementation, rather than requiring that every access of a BINARY!
needs to weigh in mind whether its underlying data is external or not.

To support this feature, the ability to store two pointers in a HANDLE!
(which was already questionable) was changed to store a pointer and
a length.  That gives the intended feature--a managed node which has
data and a length but cannot be otherwise modified, and which shared
instances can mark as unavailable through the common pointer.