Simplify bootstrap by getting rid of Do_Global_Block
The Do_Global_Block routine wasn't very complicated in practice, but
had an intimidating interface with a "magic number":

    Rebind:
        0: bind set into sys or lib
       -1: bind shallow into sys (for ACTION)
        1: add new words to LIB, bind/deep to LIB
        2: add new words to SYS, bind/deep to LIB

This was both obfuscating and seemingly unnecessary (why would the
actions bind shallow into sys?)  There was also some complexity with
an "optional toplevel word" that no longer applied.

This simplifies and comments the process.  Also, it removes the
fallback for typeset creation on not requiring the words for datatypes
to be bound to the datatypes and falling back on the symbol ID, by
instead actually binding the datatypes in the specs while building the
actions and the natives.