In-place Get of variables in contexts return relative values
This is a preparatory simplifying step in order to deal with the
increase in complexity where contexts and frames store values which
may be "self-relative".  This property means that copying variables
out of contexts to other locations will have to be handled as with
other arrays, but where the specifier to use is the object or frame
itself.

(The goal of the coming change is to remove the requirement of frame
reification for user functions unless a self-relative value--such as
a definitional RETURN or a VARARGS!--winds up being copied to a cell
that does not live in the frame.)

Since several call sites for variable fetching were touched in this
commit, it also gets rid of the all-caps for the variable fetching
routine names.