Clean up function meta object code
Functions were changed to not preserve their spec block, rather they
were able to preserve more structural information as a "meta" object
that could be queried by field.  The prototypes for the meta objects
were included in %sysobj.r.

There seemed to be a misunderstanding about the need to do SPECIALIZE
during bootstrap before the sysobj.r was available, resulting in the
specialized meta object being built by hand in C code.  There was also
no use of the auto-generated indexes of fields for system objects,
which is one of the main features.

This corrects the misunderstanding and uses the field indices.  It
also makes the object that does have to be built by hand (the ordinary
function meta information, built for natives and actions) correspond
directly to the indices, by becoming "selfish".