Fix mezz-legacy.r BREAK
Previously, the /RETURN form of BREAK was supported even if not in
legacy mode.  With the new rules, /RETURN has been dropped altogether
because BREAK forces the loop to return a blank.  So the message
explaining why this is simply not supported is given.

In any case, the movement of BREAK into `<r3-legacy>` was done
incorrectly.  This fixes it.