REDUCE leaves BAR! in blocks, SET and GET tolerate BAR!
The evaluator skips over BAR!s, so that DO/NEXT does not see them.
They are considered conceptually void and "opt-out", so that
**do [| | |]** is a void, **all [true | true]** is true and,
**any [false | false]** is false.

This raises the question of how they should be treated by a REDUCE
operation.  Originally, since a REDUCE was merely moving along in
DO/NEXT steps, they would be omitted from the result so that
**reduce [1 + 2 | 3 + 4]** would be **[3 7]**.  This change makes the
REDUCE itself preserve BAR!s, so it would be **[3 | 7]**.

In order to allow BAR!s to provide a structural check on SET operations
this extends SET to allow a BAR! in a position so long as it lines up
in the right position.  Hence you can write:

    set [a | b] [1 + 2 | 3 + 4]

For compatibility, GET is changed to allow blocks as well, with BAR!s
at the positions:

    a: 1
    b: 2
    get [a | b] ;-- would be [1 | 2]

This change also lets you use BLANK! to skip assigning some results,
with the following assigning 1 to a and 3 to b, but throwing away 2:

    set [a _ b] [1 2 3]