Reworked extension mechanism
Instead of having commands in the extensions, new extensions contain
natives, just like the core.

The script make-ext-natives.r scans the c source code, and extracts the
natives. It works in a very similar way as make-natives.r for the core
natives, just with some enhancements:

1. native word in the spec supports /export to signify that the native
will be exported, when the module is loaded.
2. if there is a 'new-words:' immediately following the native spec, the
words will be collected as the REBOL words that natives will need. The
canonical form of the words will be saved, and a
(module-name)_WORD_(word-name) will be defined for each word.

for example, in n-crypt.c:
//
//  rsa: native/export [
//
//  "Encrypt/decrypt data using the RSA algorithm."
//
//      data [binary!]
//      key-object [object!]
//      /decrypt
//         "Decrypts the data (default is to encrypt)"
//      /private
//         "Uses an RSA private key (default is a public key)"
//      /padding
//          "Selects the type of padding to use"
//      padding-type [word! blank!]
//          "Type of padding. Available values: PKCS1 or NONE"
//  ]
//  new-words: [n e d p q dp dq qinv pkcs1]
//

"rsa" will be exported to the lib and user context when the module
"crypt" is loaded, because it's exported (specified by native/export)

Macros for the canonical forms of words in [n e d p q dp dq qinv pkcs1]
will be created, e.g. CRYPT_WORD_N for "n".

make-ext-natives.r produces three files:
1. tmp-ext-(module)-args.h for macros like INCLUDE_PARAMS_OF_(native)
2. tmp-ext-(module)-words.h for macros like (moudle)_WORD_(word)
3. tmp-ext-(module).h
    This file has to be included after definitions of natives, because
    it references them. It contains the specs of all natives in the
    file, an array of all C function pointers to the implementation
    of the natives, and a pair of function for module initialization and
    deinitialization.

To avoid name conflicts, the natives in an extension should be declared
with "static".

Extension (collection of modules) loading process

Currently only builtin extensions are supported. On C side: An module
can be added to exe by calling LOAD_MODULE(module), what it really does
is saving the spec and array of C functions (converted to rebol handle!)
to boot-exts. On Rebol side, when Rebol boots, host-start.r will run,
which calls load-boot-exts, which then reads boot-exts and creates
natives with the spec and C function pointer (by load-native), and
finally makes a module. It will export functions in the module if there
are any.

An implementation detail: The module contains a hidden member "words"
that is a block of words collected from "new-words" in the C source
file. They are kept here to make sure that words will stay and their
canonical forms in Ext_(module)_Canons won't be invalided by GC.