Switch re-use of varargs to separate REB_0_PICKUP type
During PATH! processing, refinements are pushed to the stack as WORD!.
Then they are popped off the stack as they are processed.

However, it is possible for the user to invoke a refinement out of
order from the order of definition in the function.  This means that
a used refinement might be visited "before its time"--e.g. before it
should consider the next expression at the callsite to be its
argument (e.g. it is for another refinement defined later).

In these circumstances, the WORD! was mutated into a VARARGS!, because
a varargs had the same kind of data payload that was needed.  However,
this was a fairly incidental sharing that limited the adaptation of
VARARGS!.

Since the type is a transitional one only used internally to the
evaluator, the REB_0 was reused to make a Reb_Pickup payload to hold
the position of the parameter and argument.  This lets the definition
of VARARGS! be changed without affecting refinement processing in
the core evaluator.