Allow extensions to define their own errors
new errors are specifiedy by "new-errors:" block after the native
spec. A error number will be created for each error in the form of:
RE_EXT_(module)_(key), and they will be unique through the runtime of
the interpreter, because each module will have a different error-base
that is initialized when the module is loaded. The error-base and error
specs are passed to load-boot-modules through the "boot-exts" block, and
proper error object was then inserted into system/catalog/errors.

To avoid too many generated header files for each module, they are
reduce to two files: tmp-ext-(module)-first. and -last.h. The former
should be included before the first native is defined, and the latter
one should be included after the last native is defined.