Check context type in a handle before using it
Because the context is passed to user script as a handle, different
handles could be mixed up, they have to be verified when passed back to
C, or bad things will happen.

Using the cleanup function as the identity of the handle as @Hostilfork
suggested.

so this reports an error as it should
>> k: make binary! 16
>> insert/dup k #{ab} 16
>> aes-ctx: aes/key k _
>> rc4/stream aes-ctx #{abcd}
** Crypt error: Not a RC4 context: handle!
** Where: rc4
** Near: rc4/stream aes-ctx #{ABCD} ??