Fix RETURN in CHAIN, remove EXIT=QUIT, leaked series on QUIT
This addresses a few issues.  One is that EXIT (which is used to
implement definitional returns) was made to exit if you used it at the
top level (so at the REPL it would act identically to QUIT).  But it
wasn't checking to see if it was a definitional exit that didn't get
caught by its proper function.  Resolved this just by getting rid of
the idea that EXIT quits the interpreter--that just confuses matters.

Another is that the startup code binary was leaked if there was a QUIT
(or EXIT quit equivalent) at a certain time.  Simple issue, fixed.

The trickier issue is that when a definitional RETURN was being loaded
into a frame slot, it was bound relative to the outer function and
not the underlying function.  However, it is the underlying function
which will be in the f->func slot at the time code that can run a
RETURN actually runs.  This changes it to use the underlying function,
which isn't a perfectly generalized fix to the concern, but it makes
CHAIN work--at least.