Work around apparent optimization bug in GCC 5.4 -O2
This tweaks an assert from a chain of boolean || into a chain of IFs,
which for some reason bypasses a bad code generation seen in GCC 5.4.
In examining the assembly, this appears to be a compiler bug--at least
undefined behavior sanitizer doesn't seem to offer any clues as to
what could be causing the problem.

Also fixes a warning in the C++ build of the new extension model, and
uses Init_Endlike_Header for the END_CELL.