Change it so FUNCTION! uses slightly more typical path dispatch
In R3-Alpha, path dispatch was done by type-specific hooks, but there
was no dispatcher for FUNCTION! itself.  Instead, the NULL dispatcher
on functions halted path processing, and the remainder was processed
as a list of refinements.

Conceptually, Ren-C wants to think of APPEND/ONLY/DUP as an "ordinary"
path dispatch, where (APPEND/ONLY) creates a "refined" function which
can be further refined by (APPEND/ONLY)/DUP.  As an efficiency trick,
the stack is used to collect refinements in the case that the product
of the path is intended to be executed.

This change makes the trick part of path dispatch itself, so that
there is an actual PD_Function running to process refinements.