Make void illegal in SELECT and path selection
Previously void was being used to "opt out" of a function refinement.
This made it easy to write:

    append/(if condition ['only]) blk data

This didn't seem very invasive because functions did not use ordinary
path dispatch.  However, it seems generally unsafe to ignore voids
in path processing and go on to the next path item.  Using blanks is
not all that much more code:

    append/(all [condition 'only]) blk data

However, the interpretation of blanks in path selection needs to be
type specific, because they are legal "values" and hence legal anywhere
an ANY-VALUE! is legal, such as a key in a MAP!

    m: make map! [_ 10]
    m/_ ;-- should be 10

This keeps the interpretation of BLANK! as being up to the type, and
makes void trigger an error in any path selection or with SELECT.