Get rid of Reb_Enumerator, simplify naming of frame functions
When frames were first being created, their invariants were just being
worked out--and the idea of "pre-fetching" was novel.  Now it is a
basic understanding of Do_Core() that it will always be called with
f.value and its evaluation type preloaded, as well as that the array
being iterated will be locked against modification.

Hence the idea of "Do_Next" and "Refetching" are intrinsically linked.
This takes the intimidating name DO_NEXT_REFETCH_MAY_THROW() and turns
it into the friendlier Do_Next_In_Frame_May_Throw().  This accompanies
other naming adjustments, such as dropping the "_MAYBE_END" from the
FETCH_NEXT_MAYBE_END() operation, as all code using it now basically
must understand that fetching (like Do-ing, or Quote-ing) must
terminate somehow.  (And alerting on END is much more robust in Ren-C
than in R3-Alpha)

In addition, there was a concept of insulating natives that ran code
from the concept of a "frame" by having a "safe enumerator", which
really was just a frame.  One idea of this enumerator was that it
might be swapped out for a less costly enumerator (like the crash-prone
"no protection" of R3-Alpha).  That idea wouldn't really work, because
all the routines being called on the enumerator are frame routines.

Since there are very slim odds that a Reb_Enumerator would ever be
anything but a frame, this helps reduce the total concept load in the
system by just getting rid of it, and using the nicer-named frame
routines instead.