Experimental VARARGS! change for more flexible MAKE VARARGS! taking
This change separates the param and arg of a VARARGS! from the data
feed sourcing it.  Each time a VARARGS! is passed as a parameter to
a variadic slot, it is kept using the same data but changed to use
the semantics of the parameter.

The idea was implemented before it was requested that there be two
forms of TAKE, e.g. TAKE or TAKE/EVAL, where TAKE/EVAL is the only
supported operation on ordinary arguments and a hard quoted argument
can do TAKE or TAKE/EVAL.  That idea is under review.

This change opens up more possible ideas in the future, by switching
to a parameter index instead of needing to store the param and arg
separately inside the varargs!.