Get rid of Fulfilling_Last_Argument logic
A complex idea was tried out which allowed operations to know when
they were fulfilling the last argument of a function, and hence it
would be possible to complete a subexpression on the left...even if
it would also be possible to continue with a lookahead operation on
the currently fulfilling argument.

This was used to try and implement an alternative scheme for "left to
right" apparent evaluation when code was all infix ops.  While it
would lead to a semantic change for the language, it was still given
consideration if it would simplify the overall picture.

However, it did not end up simplifying the picture...and in fact broke
the behavior of the new <| and |> operators (which were key motivators
for changing the code in the first place).  The opportunistic nature
of the completion led to asymmetrical and sometimes seemingly
unpredictable results.

This simplifies the logic down so that the behavior of deferment will
happen as long as any function argument is being fulfilled.