Fix direct evaluation into array in codec initialization
The evaluator does not permit using non-fixed-size array slots as the
destination for evaluations, in case something happens to resize the
array during the evaluation itself.  This applies to object varlists
as well (as they are implemented as arrays).

To catch violators, an END marker is used as the initialization bit
pattern for the output slot.  Trying to evaluate into an array slot
would thus cause an unexpected END mark at a place other than the
tail of the array.

The codec initialization was breaking this rule, and was caught by an
exercise of this check that was triggered.  This properly evaluates
the codec into a stack value cell, and then assigns it afterward.