Make error handling safe for REB_FUNCTION and REB_0_LOOKBACK
In the reporting of errors, the system climbs the stack looking for
frames to implicate.  It only wants to report those stack levels that
are running functions.

With the evaluator trying to be minimalist about what bit patterns it
pays to initialize, there was no specific pattern which could be used
to tell when a frame marked REB_FUNCTION or REB_0_LOOKBACK actually
was in the mode of being an active function call.  While several bit
patterns were guaranteed to be good if it was, nothing served as a
signal that it wasn't.

Consequently, any code that might report an error while the eval_type
was either REB_0_LOOKBACK or REB_FUNCTION (but not actually in the
midst of those operations) would temporarily set the type away to
something else, like REB_MAX_VOID.  The cases where this would pop
up as a concern were only noticed when errors occurred, so they were
worked around on an ad-hoc basis.

This change goes ahead and picks something that only the function
uses (the label) to initialize to NULL once per frame initialization,
and then back to NULL after every function finishes.  It adds a very
tiny cost but helps keep unanticipated errors from causing crashes.