Fixes so Ren-C can be used as R3-MAKE
This fixes three issues that were stopping the ability to bootstrap.

One was that LIT-PATH! parsing rules were not being supported by
Ren-C, while they had been supported by R3-Alpha.  This enables them.

Another is that the %make-ext-natives.r was using a NONE? test instead
of a BLANK? test.

Also, the REPLACE mezzanine code was switched from IF to ALL to give
a BLANK! refinement pick (since void is now illegal).  Yet this wasn't
working, because it had a refinement /ALL which was overriding that.