VARARGS! proxying is part of type checking, fix bad memory
The previous attempt to re-stamp a VARARGS! passed as a parameter
through APPLY with the new parameter signature was actually only
used in specializations, not in APPLY at all.  The problem was masked
by the use of uninitialized memory in a temporary frame's flags field
which was cooked up to allow frames to be used on MAKE VARARGS! data.