Eliminate DO_FLAG_EXECUTE_FRAME, tidy DO_FLAG_XXX
The DO_FLAG_EXECUTE_FRAME flag was set but never used; it was an
artifact which was eventually replaced by DO_FLAG_APPLYING.

This removes that flag and does some reorganization to put the "least
favorite flags" toward the end, as well as fix a typo of some
kind that was making a deliberately unused flag bit get used.  (This
would confuse the code for knowing the data type of a REBFRM*,
e.g. panic())