Replace BUF_EMIT usages with the data stack
The "emit buffer" was an array that was used primarily by the source
scanner to build GC-safe items up during an array scan, and then
they'd be copied out.  This pattern of usage was completely redundant
with the data stack.

The data stack is more optimized and cleaner to use with the pattern.
It doesn't need to update its size with each push and pop, and it is
checked for balance.