Restore expansion safety for MAKE inside of a Scan_Array
Eliminating the emit buffer and doing scans directly into a value cell
on the data stack also (mistakenly) eliminated the safety of running
MAKE (construction syntax requiring execution) during a scan into
a "stable" cell location.  This puts that back, because the same rule
applies to the data stack as would apply to the emit buffer.

In addition, this adds warning to the relatively simple Scan_XXX
routines (e.g. Scan_Integer) to not call into the evaluator, GC, or
use DS_PUSH.  It is only the last requirement that is new; so the
routines were already supposed to be somewhat "atomic".