Make STRUCT! inaccessibility come from handle size = 0
Originally, the series flag for inaccessibility originated from the
FFI code.  To save on space, it was co-opted by the code for FRAME!
series, to indicate when a frame was no longer on the stack.

The FFI struct feature had another flag, SERIES_FLAG_EXTERNAL.  This
was replaced in a previous commit by the idea that a struct could be
backed by either a binary series, or a singular array series containing
a HANDLE! value.  This let the "external" property be contained to the
FFI's specific application, while not requiring the entire system to
be bulletproofed against the concerns of arbitrary series which might
not have their memory under control of Rebol (which would be numerous).

Though SERIES_FLAG_EXTERNAL was replaced, the inaccessibility flag was
kept between both frames and HANDLE!-backed structs.  One reason for
using the flag instead of just overwriting the HANDLE! to NULL was
because error messages wanted to deliver what the inaccessible
pointer value had been.

Yet this gave very different meanings to the flag, making managing the
concerns between frames and the FFI complicated.  A simpler solution
that can still keep the pointer for the error messages is just to
indicate that a HANDLE!-backed struct is inaccessible is to set the
length in the handle to zero.

This also adds back in typechecking to the ARR_XXX() routines, which
due to the generality of AS_SERIES() had lost that check.  Adds code
to handle PROBE(context).