Separate var inaccessibility from REBFRM* availability of FRAME!
Two properties had been twisted together incorrectly for a FRAME!,
which were whether or not that frame had code currently running on
the stack corresponding to it, and whether the variables in the
frame were accessible.  In `<durable>` (a.k.a. "CLOSURE!") semantics,
it would be possible for a function to be finished on the stack and
yet its variables were still available.

This reviews and tightens up the properties, reflecting that the
REBFRM* is stored in the series ->misc field itself.  (Originally it
was stored in the REBVAL* for a frame, which led to the idea that it
could not be examined for frames that weren't on the stack.)