Add RECYCLE/VERBOSE to Debug build
This is an experimental basic way of getting some information about
the nodes which are being garbage collected.  Using a manually
allocated array from before the recycle is run, it gathers the
nodes expected to be freed.  The nodes are then PROBE'd one at a time,
and then the actual recycling process is performed.

The code for filling the "sweeplist" array is separated from the
actual sweeping code, to avoid slowing it down with debug tests.

(Includes a minor fix to VARARGS! on expired frames, an added test
in ASSERT_CONTEXT(), and removal of the recursion-causing PROBE()
when panic'ing on a series)