Implement LOCK thru PROTECT
Both FROZEN and PROTECT flags will be set when a value is LOCKed,
however, UNPROTECT will refuse to unprotect a series/context if the
FROZEN flag is set, or the containing context of a word is FROZEN, thus,
the protection becomes permanent (no way to unprotect a locked value).

When the value is to be modified, only the PROTECT flag is checked, and
an approriate error is raised if it's read-only.

This also changed the error key from protected-word to word-protected to
make it consistent with series-protected and context-protected.