Fix LOCK of objects and native frames + PROTECT of individual keys
This reverts the commit which used per-key settings to LOCK objects,
and instead makes basic PROTECT and LOCK use the same method of
just setting bits on the varlist series.  Setting variables and
objects through paths is changed to honor that those bits.

The per-key PROTECT feature is slightly different, as it requires any
object to which this is applied to have a unique keylist.  This does
the necessary ensuring that a field's keylist is unique (may require
making a duplicate) before modifying the bits of the key.

Because variable access is now honoring the protection status of
objects instead of just keys, the changes meant that a FRAME! made for
any "non-interpreted" functions was being treated as locked.  This
is a problem for ADAPT, because it wants to be able to run code that
modifies frame values prior to a native it adapts running...yet once
that native is running, a FRAME! tied to the call should not be able
to modify the native's frame values out from under the C code (which
could lead to a crash).  This introduces a basic workaround to the
problem where once native code starts running on a frame, that is
when the frame is marked read-only to any outside access (e.g. not
through FRM_ARG())