Make PIPE()-based CALL work asynchronously without PIPE2()
NOTE: pipe() is POSIX, but pipe2() is Linux-specific.  With pipe() it
takes an additional call to fcntl() to request non-blocking behavior,
so it's a small amount more work.  However, there are other flags which
if aren't passed atomically at the moment of opening allow for a race
condition in threading if split, e.g. FD_CLOEXEC.

(If you don't have FD_CLOEXEC set on the file descriptor, then all
instances of CALL will act as a /WAIT.)

At time of writing, this is mostly academic...but the code needed to be
patched to work with pipe() since some older libcs do not have pipe2().
So the ability to target both are kept around, saving the pipe2() call
for later Linuxes known to have it (and O_CLOEXEC).