If arguments aren't evaluating, allow BAR!
Argument processing for cases like Do_Va or APPLY were not allowing
a BAR! to be preloaded into an argument slot.  This was exposed by
the console code, which tried to pass the result of previous
evaluations into a function called from C.

Also fixes PRIN to respect the new rule that voids are not allowed
in path dispatch.