Reduce dependence on PRINT in %host-main, Ctrl-C bugfix
This adds some additional debug information for DROP_GUARD_SERIES
and DROP_GUARD_VALUE so that it's possible to at least know where the
failure came from.  Using this, a bug in Ctrl-C handling was found
when the REPL is interrupted while doing multi-line input.

Also simplifies the host's responsibilities by disallowing the normal
PRINT-ing of any errors while running the startup code.  A normal
error scenario in the startup should be TRAP'd and printed by that
code, while it will be considered a panic-able condition by the C
shell.  This division of labor helps avoid the need to manage the
semantics of printing logic (or interruptions of that print process,
e.g. via Ctrl-C on a long error printout) from C code.