Fix spurious VALUE_FLAG_LINE in %l-scan.c
An assertion discovered that end markers were having VALUE_FLAG_LINE
set on them, which is not legal.  But moving it to the last value did
not have the correct effect, since the intent was that the newline is
*after* the last value...this was causing formatting weirdness.

It's not clear exactly what's broken by virtue of dropping the line
indicator on the floor--things appear to work normally.