Automatically add backtraces in error generating functions
SET-WHERE-OF-ERROR was added as a native to complement DO of an ERROR!
as the only other mechanic required to write FAIL as user code.

The function for manipulating the backtrace of an error was called
"Try_Add_Backtrace_May_Fail", which as the name suggests was not very
rigorous in terms of how the backtrace was calculated.  The idea of
making more callers responsible for whether it was an appropriate
moment to add a backtrace or not was tested.

This led to some code paths for error creation not adding a backtrace,
since some would crash if they did try.  This puts the responsibility
for "guessing" on the caller of the backtrace adding, but goes back
to putting that guess as part of all error creation routines...as
opposed to in the backtrace