LOCK code which is not in modules, run from the command line
The code path for plain files that are not modules was not locking
source, as it was being locked in the console.  This meant code that
assumed source had to be locked (namely using source as a key in a
MAP!) couldn't operate under that expectation.

Also, though it is not currently enforced, the REPL was locking code
before it was binding it.  This moves the lock to after the bind.

(Changing the binding will be considered a modification of locked
source, but this doesn't have an error yet because the workaround for
getting a copy of locked source with different binding that doesn't
actually involve copying the block itself is done with
"virtual binding", which hasn't been finished yet.)