Eliminate redundant DO code in usermode startup function
There was a sort of impromptu rewrite of DO of a FILE! for running
programs from the command line.  The comment from R3-Alpha said:

    ; !!! Would be nice to use DO for this section. !!!
    ; NOTE: We can't use DO here because it calls the code it does
    ; with CATCH/quit and we shouldn't catch QUIT in the top-level
    ; script, we should just quit.

The redundancy of code was a very bad thing.  This corrects the issue
by adding a refinement to DO to tell it not to catch quits, even for
a file.  (It's named /ONLY, which might be a bad choice, but it will
do for now.)

Should resolve #382.