Move Push_Frame and Drop_Frame into callers of Do_Core
Unless an apply was running, the code in Do_Core would push the frame
onto the stack and drop it.  This was fine for DO of a block, since
the evaluator loop would run over the whole block.  But for commands
like ANY or ALL which set up a frame and then manage the logic of
each successive evaluation, it meant pushing the frame and then
dropping it on each call to Do_Core.  It even meant the same frame
could be pushed to the stack twice.

This addresses the issue by requiring the caller to do the frame
push, and then Do_Core() asserts it is only acting on the topmost
frame in the stack.