Add experimental DO-ALL native
This adds a native which runs several spans of code (separated by BAR!
in a block).  It was designed to try and address error propagation in
the sys/do* helper, but then not used for the purpose.  However, this
code or something similar should exist.

Hence this is committing the work, just to keep track of it.