Make seeks in PARSE via GET-WORD! reset `begin` between rules
Cryptically, the INSERT command does not use the "current" parse
position, but rather the "begin" position...where the parse position
was when the current rule started.  It's not clear why that was chosen
though it looks to have been on purpose.

This leads to quirky behavior when coordinating with GET-WORD! seeks
that change the parse position, because those seeks never actually
reset the `begin` point.  This ties into a broader issue discussed in
an issue about what MARK and SEEK mean when used while a partially
expressed rule is in progress:

https://github.com/rebol/rebol-issues/issues/2269

As a minimum disruption to at least fix the most obvious problem for
now, this makes any seeks that aren't "in the middle of a rule"
(e.g. flags=0 due to post_rule_processing) reset the begin position,
as if they were their own rule.

#2269 is left open for discussion.