Add experimental SET-UDP-MULTICAST and SET-UDP-TTL natives
This adds natives which intend to be equivalent to the Rebol2 calls:

    set-modes port [multicast-groups: [[(...) (...)]]]
    set-modes port [multicast-ttl: (...)]

The code is based on observations and findings discussed here, and
it seems reasonable to believe it would work:

https://github.com/zsx/r3/issues/38

Having it work is contingent upon UDP receiving ports working, which
would need to be demonstrated.

In order to pass parameters through from the core to the OS-specific
network code, this embraces the direction of not prohibiting "host"
or "extension" code from being able to call the implementation
functions of the interpreter.  This means that port files such as
%p-net.c could hopefully be moved into their own extension as well,
and merged with the "device" code that they are artificially separated
from at this time.