Eliminate unused variable-sized REBREQ "feature"
The "Device" layer in R3-Alpha was not intended to be able to use
features from the closed-source Rebol core.  This meant it was
necessary to translate from higher-level Rebol types and abilities
down to the devices, even though the devices were authored by the
same people who had source code to Rebol--and the "Device" spec did
not represent the end of the line, since it mapped to no hardened
and unchangeable OS API.

This means that the device layer wound up trying to be a user of C
tricks, such as the trick of allowing a variable-sized structure to
be overlaid onto a fixed-size C structure header by prefixing it with
an actual length.  This was theoretically something that devices could
use to have REBREQs of larger size.

The feature was not used (all sizes were 0, e.g. default to
sizeof(REBREQ), or literally sizeof(REBREQ)).  It also represents a
poor axis of extensibility when Rebol arrays and objects are available.

This is a first step toward trying to migrate requests to accept
Rebol types and values (e.g. STRING!) instead of having an OS_ALLOC
and OS_FREE of REBCHR to put in the request originating from the core.